+++ packages/cli/package.json
@@ -32,11 +32,13 @@
     "semver": "^7.3.5",
     "winston": "~3.8.2",
     "diod": "~2.0.0",
-    "ts-pattern": "~4.2.1"
+    "ts-pattern": "~4.2.1",
+    "@oclif/help": "~1.0.5"
   },
   "devDependencies": {
     "@boostercloud/eslint-config": "workspace:^1.7.0",
     "@boostercloud/application-tester": "workspace:^1.7.0",
+    "@boostercloud/metadata-booster": "workspace:^1.7.0",
     "@oclif/dev-cli": "^1.26",
     "@oclif/test": "^2.1",
     "@types/chai": "4.2.18",
@@ -70,8 +72,10 @@
     "rimraf": "3.0.2",
     "sinon": "9.2.3",
     "sinon-chai": "3.5.0",
+    "reflect-metadata": "0.1.13",
     "ts-node": "^10.9.1",
     "typescript": "4.7.4",
+    "ttypescript": "1.5.13",
     "eslint-plugin-unicorn": "~44.0.2",
     "@types/semver": "5.5.0"
   },
@@ -101,9 +105,9 @@
     "format": "prettier --write --ext '.js,.ts' **/*.ts **/*/*.ts",
     "lint:check": "eslint --ext '.js,.ts' **/*.ts",
     "lint:fix": "eslint --quiet --fix --ext '.js,.ts' **/*.ts",
-    "build": "tsc -b tsconfig.json && copyfiles -f src/templates/*.stub dist/templates",
+    "build": "ttsc -b tsconfig.json && copyfiles -f src/templates/*.stub dist/templates",
     "clean": "rimraf ./dist tsconfig.tsbuildinfo",
-    "prepack": "tsc -b tsconfig.json",
+    "prepack": "ttsc -b tsconfig.json",
     "test:cli": "npm run test",
     "test": "nyc --extension .ts mocha --forbid-only  \"test/**/*.test.ts\""
   },
