+++ packages/cli/src/commands/add/projection.ts
@@ -9,8 +9,8 @@ import {
 } from '../../services/file-generator/target'
 import Brand from '../../common/brand'
 import { generateProjection, getResourceSourceFile } from '../../services/method-generator'
-import { Logger } from 'framework-types/dist'
-import { UserProject } from 'cli/src/services/user-project'
+import { Logger } from '@boostercloud/framework-types'
+import { UserProject } from '../../services/user-project'
 
 @CliCommand()
 class Implementation {
