+++ packages/cli/src/commands/new/project.ts
@@ -1,10 +1,10 @@
 import { flags } from '@oclif/command'
 import { IConfig } from '@oclif/config'
-import { Args, BaseCommand, CliCommand, Flags } from 'cli/src/common/base-command'
-import { FileGenerator } from 'cli/src/services/file-generator'
-import { UserInput } from 'cli/src/services/user-input'
-import { ProjectCreationConfig, UserProject } from 'cli/src/services/user-project'
-import { Logger } from 'framework-types/dist'
+import { Args, BaseCommand, CliCommand, Flags } from '../../common/base-command'
+import { FileGenerator } from '../../services/file-generator'
+import { UserInput } from '../../services/user-input'
+import { ProjectCreationConfig, UserProject } from '../../services/user-project'
+import { Logger } from '@boostercloud/framework-types'
 import Brand from '../../common/brand'
 import { Provider } from '../../common/provider'
