+++ packages/cli/src/commands/new/read-model.ts
@@ -13,9 +13,9 @@ import {
 } from '../../services/file-generator/target'
 import * as path from 'path'
 import { classNameToFileName } from '../../common/filenames'
-import { Logger } from 'framework-types/dist'
-import { FileGenerator } from 'cli/src/services/file-generator'
-import { UserProject } from 'cli/src/services/user-project'
+import { Logger } from '@boostercloud/framework-types'
+import { FileGenerator } from '../../services/file-generator'
+import { UserProject } from '../../services/user-project'
 
 export default class ReadModel extends BaseCommand<typeof ReadModel> {
   public static description = 'create a new read model'
