+++ packages/cli/src/commands/nuke.ts
@@ -2,9 +2,9 @@ import { flags } from '@oclif/command'
 import { BaseCommand, CliCommand, Flags } from '../common/base-command'
 import { Logger } from '@boostercloud/framework-types'
 import Brand from '../common/brand'
-import { UserInput } from 'cli/src/services/user-input'
-import { UserProject } from 'cli/src/services/user-project'
-import { CloudProvider } from 'cli/src/services/cloud-provider'
+import { UserInput } from '../services/user-input'
+import { UserProject } from '../services/user-project'
+import { CloudProvider } from '../services/cloud-provider'
 
 @CliCommand()
 class Implementation {
