+++ packages/cli/src/index.ts
@@ -1,1 +1,2 @@
+import 'reflect-metadata'
 export { run } from '@oclif/command'
