+++ packages/cli/src/services/cloud-provider/generic.impl.ts
@@ -46,7 +46,6 @@ export class GenericCloudProvider implements CloudProvider {
     methodName: TMethodName
   ): Promise<NonNullable<ProviderInfrastructure[TMethodName]>> {
     const config = await this.userProject.loadConfig()
-    await this.assertNameIsCorrect(config.appName)
 
     const method = config.provider.infrastructure()[methodName]
     if (!method) {
@@ -56,34 +55,5 @@ export class GenericCloudProvider implements CloudProvider {
       )
     }
     return method as NonNullable<ProviderInfrastructure[TMethodName]>
-  }
-
-  async assertNameIsCorrect(name: string): Promise<void> {
-    type StringPredicate = (x: string) => boolean
-
-    // Current characters max length: 37
-    // Lambda name limit is 64 characters
-    // `-subscriptions-notifier` lambda is 23 characters
-    // `-app` prefix is added to application stack
-    // which is 64 - 23 - 4 = 37
-    const maxProjectNameLength = 37
-
-    const constraints: ReadonlyArray<[StringPredicate, string]> = [
-      [(name: string) => name.length > maxProjectNameLength, `be longer than ${maxProjectNameLength} characters`],
-      [(name: string) => name.includes(' '), 'contain spaces'],
-      [(name: string) => name.toLowerCase() !== name, 'contain uppercase letters'],
-      [(name: string) => name.includes('_'), 'contain underscore'],
-    ]
-
-    for (const [constraint, restrictionText] of constraints) {
-      if (constraint(name)) {
-        const message = this.nameFormatErrorMessage(name, restrictionText)
-        throw new CliError('CloudProviderError', message)
-      }
-    }
-  }
-
-  private nameFormatErrorMessage(name: string, restrictionText: string): string {
-    return `Project name cannot ${restrictionText}:\n\n    Found: '${name}'`
   }
 }
